+++ src/lib/Scrolly/Scrolly.stories.svelte
@@ -31,13 +31,13 @@
 </script>
 
 <script>
-  import { useScrollyState } from "$lib";
+  import { useScrollyContext } from "$lib/context";
 </script>
 
 {#snippet template(args)}
   <Scrolly {...args}>
     {#snippet background()}
-      {@const scrolly = useScrollyState()}
+      {@const scrolly = useScrollyContext()}
       {@const totalSlides = args.slides?.length ?? 0}
       <div
         style="width: 100%; height: 100vh; color: var(--color-white); font-weight: var(--font-weight-bold); background: var(--color-blue); display: flex; flex-direction: column; gap: var(--spacing-4); align-items: center; justify-content: center;"