+++ src/lib/index.js
@@ -33,4 +33,4 @@ export { default as LoadingWrapper } from "./LoadingWrapper/LoadingWrapper.svelt
 export { default as Tooltip } from "./Tooltip/Tooltip.svelte";
 export { default as AxisY } from "./Layercake/AxisY.svelte";
 export { default as AxisX } from "./Layercake/AxisX.svelte";
-export { useScrollyState } from "./Scrolly/state.svelte.js";
+export { useScrollyContext } from "./Scrolly/context.svelte.js";