+++ src/lib/maps/SVGPointLayer/SVGPointLayer.stories.svelte
@@ -31,13 +31,13 @@
 <script>
   import states from "../../../docs/sample-data/states_geo.json";
   import { urbanColors } from "$lib/utils";
-  import { createMatchMedia } from "$lib/stores";
+  import { createMatchMediaContext } from "$lib/context";
 
   let mousemoveHandler = fn();
   let mouseoutHandler = fn();
   let clickHandler = fn();
 
-  createMatchMedia();
+  createMatchMediaContext();
 </script>
 
 <Story